+++ test/solidityTypes.ts
@@ -5,6 +5,6 @@ import { BytesLike } from '@ethersproject/bytes'
 export type address = string
 export type uint256 = BigNumberish
 export type uint = BigNumberish
-export type uint64 = BigNumberish
+export type uint48 = BigNumberish
 export type bytes = BytesLike
 export type bytes32 = BytesLike
